requirements.txt
feedparser>=6.0.11
numpy>=1.24.4
pillow>=10.2.0
qrcode>=7.4.2